+++ in-adapter-api/src/main/java/gohigher/config/SpringSecurityConfig.java
@@ -11,8 +11,8 @@ import org.springframework.security.config.annotation.web.configurers.CsrfConfig
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 
-import gohigher.MyAuthenticationSuccessHandler;
-import gohigher.OauthUserService;
+import gohigher.oauth2.MyAuthenticationSuccessHandler;
+import gohigher.oauth2.OauthUserService;
 import lombok.AllArgsConstructor;
 
 @Configuration
