+++ in-adapter-api/src/main/java/gohigher/oauth2/MyAuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package gohigher;
+package gohigher.oauth2;
 
 import java.io.IOException;
