+++ in-adapter-api/src/main/java/gohigher/oauth2/OAuth2Attribute.java
@@ -1,4 +1,4 @@
-package gohigher;
+package gohigher.oauth2;
 
 import java.util.HashMap;
 import java.util.Map;
