+++ in-adapter-api/src/main/java/gohigher/oauth2/OauthUserService.java
@@ -1,4 +1,4 @@
-package gohigher;
+package gohigher.oauth2;
 
 import java.util.Collections;
 import java.util.Map;
